Early Camera Implementation
Scenes now have a a propery .camera which holds an object with
functional properties .x and .y.  Angle and zoom are not yet
implemented.

Layers now also have .scrollRateX and .scrollRateY properties, which
affect how much a layer appears to move when the camera is moved.  For
example, if a layer's scrollRateX is set to 0.5, objects on this layer
will appear to move at half the speed of the camera.